pipeline/publication/requirements.txt
PyYAML==6.0.1
wandb==0.16.0
